requirements.txt
Django==2.2.12
django-bootstrap4==1.1.1
django-decouple==2.0
django-model-utils==4.0.0
django-simple-history==2.10.0
psycopg2-binary==2.8.5
pytz==2020.1
sqlparse==0.3.1
whitenoise==5.0.1
